This commit adds support for "version" in the OpenTSDB protocol.
tcollector uses 'version' to determine if an OpenTSDB endpoint is up.  It expects the server to respond with something if it sends "version" (it doesn't matter what - but it expects the response to have something)
See the tcollector.py for more details:
https://github.com/OpenTSDB/tcollector/blob/master/tcollector.py#L581

Without this change, tcollector will put the Wavefront endpoing into its blacklist and not send it any more metrics.