Create new "nodeps" jar as part of proxy release process.
Our maven-assembly-plugin replaces the common type of jar (no
dependencies) with an uber jar, since we have appendAssemblyId=false.
maven-assembly-plugin would support output naming like this:

        proxy-3.23.jar  // Small, no deps.
        proxy-3.23-with-deps.jar  // Large, uber.

Our current scheme is this:

        proxy-3.23.jar  // Large, uber.

This commit should bring us to this scheme:

        proxy-3.23.jar  // Large, uber.
        proxy-3.23-nodeps.jar  // Small, no deps.

This is tested using the
-DaltDeploymentRepository=snapshots::default::file:///myFile flag for
the deploy goal of the deploy plugin.