Throw explicit errors on dupe metrics.
Because metric names can be dynamic
(i.e. "my_metric_%{field_from_logs}"), I can't know if a user is trying
to use the same metric name to refer to differing metric types. Now, in
this (hopefully rare) case, an exception is thrown and the new metric
will not be parsed.