Update README.md
Edited existing sections, and added a new "Set Up a Wavefront Proxy" section that can be linked to from the READMEs for the new observability SDKs. This section provides a streamlined set of steps for the SDK user, and might eventually be moved into the docs. For now, it enables the SDK README's to link to the information without waiting for 29.x docs to go live.