PUB-218 enhanced preprocessor rules. (#509)
* PUB-218 enhanced preprocessor rules.

Gist:
We introduced a new argument denoted in yaml with the key "if", whose value denotes a predicate function. The preprocessor rule will apply to a given entity if this predicate function evaluates to true. We will be calling this as the "v2PredicateKey" and its corresponding predicate function as "v2Predicate".

Semantics and behavior changes:
1. v2Predicate applies to all scopes except for "pointline"
2. For "Filtering" preprocessor rules, Previously v1Predicates(i.e. [match], [scope] arguments) were considered mandatory. With this change, v1Predicates are considered optional, if v2Predicate present. However, note that v1Predicates cannot be specified and left blank(this is still considered invalid). If both v1Predicates and v2Predicate is specified, then both need to be satisfied to apply the rule.